investra-gov-apps-be/requirements.txt
Flask==3.1.*
Flask-SQLAlchemy==3.1.*
Flask-Migrate==4.1.*
Flask-CORS==5.0.*
Flask-Limiter==3.8.*
redis==5.2.*
psycopg2-binary==2.9.*
scikit-learn==1.6.*
numpy==2.2.*
pandas==2.2.*
python-dotenv==1.1.*
gunicorn==23.0.*
PyJWT==2.9.*
Werkzeug>=3.0
